requirements.txt
google-generativeai>=0.7.0
fastapi>=0.111.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
google-genai

google-generativeai

fastapi 
uvicorn 
pandas 
sqlite3 
requests 
python-multipart